break sync, interop, and value code into their own files
