+++ automerge-wasm/src/lib.rs
@@ -1,13 +1,19 @@
-extern crate web_sys;
 use automerge as am;
-use automerge::{Change, ChangeHash, ObjId, Prop, Value, ROOT};
-use js_sys::{Array, Object, Reflect, Uint8Array};
-use std::collections::{HashMap, HashSet};
-use std::convert::TryFrom;
+use automerge::{Change, ObjId, Prop, Value, ROOT};
+use js_sys::{Array, Object, Uint8Array};
 use std::convert::TryInto;
-use std::fmt::Display;
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
+
+mod interop;
+mod sync;
+mod value;
+
+use interop::{
+    get_heads, js_get, js_set, map_to_js, to_js_err, to_objtype, to_prop, to_usize, AR, JS,
+};
+use sync::SyncState;
+use value::{datatype, ScalarValue};
 
 #[allow(unused_macros)]
 macro_rules! log {
@@ -20,101 +26,9 @@ macro_rules! log {
 #[global_allocator]
 static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
 
-fn datatype(s: &am::ScalarValue) -> String {
-    match s {
-        am::ScalarValue::Bytes(_) => "bytes".into(),
-        am::ScalarValue::Str(_) => "str".into(),
-        am::ScalarValue::Int(_) => "int".into(),
-        am::ScalarValue::Uint(_) => "uint".into(),
-        am::ScalarValue::F64(_) => "f64".into(),
-        am::ScalarValue::Counter(_) => "counter".into(),
-        am::ScalarValue::Timestamp(_) => "timestamp".into(),
-        am::ScalarValue::Boolean(_) => "boolean".into(),
-        am::ScalarValue::Null => "null".into(),
-    }
-}
-
-#[derive(Debug)]
-pub struct ScalarValue(am::ScalarValue);
-
-impl From<ScalarValue> for JsValue {
-    fn from(val: ScalarValue) -> Self {
-        match &val.0 {
-            am::ScalarValue::Bytes(v) => Uint8Array::from(v.as_slice()).into(),
-            am::ScalarValue::Str(v) => v.to_string().into(),
-            am::ScalarValue::Int(v) => (*v as f64).into(),
-            am::ScalarValue::Uint(v) => (*v as f64).into(),
-            am::ScalarValue::F64(v) => (*v).into(),
-            am::ScalarValue::Counter(v) => (*v as f64).into(),
-            am::ScalarValue::Timestamp(v) => js_sys::Date::new(&(*v as f64).into()).into(),
-            am::ScalarValue::Boolean(v) => (*v).into(),
-            am::ScalarValue::Null => JsValue::null(),
-        }
-    }
-}
-
 #[wasm_bindgen]
 #[derive(Debug)]
 pub struct Automerge(automerge::Automerge);
-
-#[wasm_bindgen]
-#[derive(Debug)]
-pub struct SyncState(am::SyncState);
-
-#[wasm_bindgen]
-impl SyncState {
-    #[wasm_bindgen(getter, js_name = sharedHeads)]
-    pub fn shared_heads(&self) -> JsValue {
-        AR::from(self.0.shared_heads.as_slice()).into()
-    }
-
-    #[wasm_bindgen(getter, js_name = lastSentHeads)]
-    pub fn last_sent_heads(&self) -> JsValue {
-        AR::from(self.0.last_sent_heads.as_slice()).into()
-    }
-
-    #[wasm_bindgen(setter, js_name = lastSentHeads)]
-    pub fn set_last_sent_heads(&mut self, heads: JsValue) -> Result<(), JsValue> {
-        let heads: Vec<ChangeHash> = JS(heads).try_into()?;
-        self.0.last_sent_heads = heads;
-        Ok(())
-    }
-
-    #[wasm_bindgen(setter, js_name = sentHashes)]
-    pub fn set_sent_hashes(&mut self, hashes: JsValue) -> Result<(), JsValue> {
-        let hashes_map: HashMap<ChangeHash, bool> = hashes.into_serde().map_err(to_js_err)?;
-        let hashes_set: HashSet<ChangeHash> = hashes_map.keys().cloned().collect();
-        self.0.sent_hashes = hashes_set;
-        Ok(())
-    }
-
-    fn decode(data: Uint8Array) -> Result<SyncState, JsValue> {
-        let data = data.to_vec();
-        let s = am::SyncState::decode(&data);
-        let s = s.map_err(to_js_err)?;
-        Ok(SyncState(s))
-    }
-
-    #[allow(clippy::should_implement_trait)]
-    pub fn clone(&self) -> Self {
-        SyncState(self.0.clone())
-    }
-}
-
-#[derive(Debug)]
-pub struct JsErr(String);
-
-impl From<JsErr> for JsValue {
-    fn from(err: JsErr) -> Self {
-        js_sys::Error::new(&std::format!("{}", err.0)).into()
-    }
-}
-
-impl<'a> From<&'a str> for JsErr {
-    fn from(s: &'a str) -> Self {
-        JsErr(s.to_owned())
-    }
-}
 
 #[wasm_bindgen]
 impl Automerge {
@@ -573,57 +487,6 @@ impl Automerge {
     }
 }
 
-pub fn to_usize(val: JsValue, name: &str) -> Result<usize, JsValue> {
-    match val.as_f64() {
-        Some(n) => Ok(n as usize),
-        None => Err(format!("{} must be a number", name).into()),
-    }
-}
-
-pub fn to_prop(p: JsValue) -> Result<Prop, JsValue> {
-    if let Some(s) = p.as_string() {
-        Ok(Prop::Map(s))
-    } else if let Some(n) = p.as_f64() {
-        Ok(Prop::Seq(n as usize))
-    } else {
-        Err("prop must me a string or number".into())
-    }
-}
-
-fn to_objtype(a: &JsValue) -> Option<am::ObjType> {
-    if !a.is_function() {
-        return None;
-    }
-    let f: js_sys::Function = a.clone().try_into().unwrap();
-    let f = f.to_string();
-    if f.starts_with("class MAP", 0) {
-        Some(am::ObjType::Map)
-    } else if f.starts_with("class LIST", 0) {
-        Some(am::ObjType::List)
-    } else if f.starts_with("class TEXT", 0) {
-        Some(am::ObjType::Text)
-    } else if f.starts_with("class TABLE", 0) {
-        Some(am::ObjType::Table)
-    } else {
-        None
-    }
-}
-
-struct ObjType(am::ObjType);
-
-impl TryFrom<JsValue> for ObjType {
-    type Error = JsValue;
-
-    fn try_from(val: JsValue) -> Result<Self, Self::Error> {
-        match &val.as_string() {
-            Some(o) if o == "map" => Ok(ObjType(am::ObjType::Map)),
-            Some(o) if o == "list" => Ok(ObjType(am::ObjType::List)),
-            Some(o) => Err(format!("unknown obj type {}", o).into()),
-            _ => Err("obj type must be a string".into()),
-        }
-    }
-}
-
 #[wasm_bindgen]
 pub fn init(actor: JsValue) -> Result<Automerge, JsValue> {
     console_error_panic_hook::set_once();
@@ -674,10 +537,10 @@ pub fn export_sync_state(state: SyncState) -> JsValue {
 
 #[wasm_bindgen(js_name = encodeSyncMessage)]
 pub fn encode_sync_message(message: JsValue) -> Result<Uint8Array, JsValue> {
-    let heads = get(&message, "heads")?.try_into()?;
-    let need = get(&message, "need")?.try_into()?;
-    let changes = get(&message, "changes")?.try_into()?;
-    let have = get(&message, "have")?.try_into()?;
+    let heads = js_get(&message, "heads")?.try_into()?;
+    let need = js_get(&message, "need")?.try_into()?;
+    let changes = js_get(&message, "changes")?.try_into()?;
+    let have = js_get(&message, "have")?.try_into()?;
     Ok(Uint8Array::from(
         am::SyncMessage {
             heads,
@@ -700,10 +563,10 @@ pub fn decode_sync_message(msg: Uint8Array) -> Result<JsValue, JsValue> {
     let changes = AR::from(msg.changes.as_slice());
     let have = AR::from(msg.have.as_slice());
     let obj = Object::new().into();
-    set(&obj, "heads", heads)?;
-    set(&obj, "need", need)?;
-    set(&obj, "have", have)?;
-    set(&obj, "changes", changes)?;
+    js_set(&obj, "heads", heads)?;
+    js_set(&obj, "need", need)?;
+    js_set(&obj, "have", have)?;
+    js_set(&obj, "changes", changes)?;
     Ok(obj)
 }
 
@@ -731,292 +594,3 @@ pub struct Text {}
 
 #[wasm_bindgen(js_name = TABLE)]
 pub struct Table {}
-
-fn to_js_err<T: Display>(err: T) -> JsValue {
-    js_sys::Error::new(&std::format!("{}", err)).into()
-}
-
-fn get<J: Into<JsValue>>(obj: J, prop: &str) -> Result<JS, JsValue> {
-    Ok(JS(Reflect::get(&obj.into(), &prop.into())?))
-}
-
-fn set<V: Into<JsValue>>(obj: &JsValue, prop: &str, val: V) -> Result<bool, JsValue> {
-    Reflect::set(obj, &prop.into(), &val.into())
-}
-
-struct JS(JsValue);
-struct AR(Array);
-
-impl From<AR> for JsValue {
-    fn from(ar: AR) -> Self {
-        ar.0.into()
-    }
-}
-
-impl From<JS> for JsValue {
-    fn from(js: JS) -> Self {
-        js.0
-    }
-}
-
-impl From<am::SyncState> for JS {
-    fn from(state: am::SyncState) -> Self {
-        let shared_heads: JS = state.shared_heads.into();
-        let last_sent_heads: JS = state.last_sent_heads.into();
-        let their_heads: JS = state.their_heads.into();
-        let their_need: JS = state.their_need.into();
-        let sent_hashes: JS = state.sent_hashes.into();
-        let their_have = if let Some(have) = &state.their_have {
-            JsValue::from(AR::from(have.as_slice()).0)
-        } else {
-            JsValue::null()
-        };
-        let result: JsValue = Object::new().into();
-        // we can unwrap here b/c we made the object and know its not frozen
-        Reflect::set(&result, &"sharedHeads".into(), &shared_heads.0).unwrap();
-        Reflect::set(&result, &"lastSentHeads".into(), &last_sent_heads.0).unwrap();
-        Reflect::set(&result, &"theirHeads".into(), &their_heads.0).unwrap();
-        Reflect::set(&result, &"theirNeed".into(), &their_need.0).unwrap();
-        Reflect::set(&result, &"theirHave".into(), &their_have).unwrap();
-        Reflect::set(&result, &"sentHashes".into(), &sent_hashes.0).unwrap();
-        JS(result)
-    }
-}
-
-impl From<Vec<ChangeHash>> for JS {
-    fn from(heads: Vec<ChangeHash>) -> Self {
-        let heads: Array = heads
-            .iter()
-            .map(|h| JsValue::from_str(&h.to_string()))
-            .collect();
-        JS(heads.into())
-    }
-}
-
-impl From<HashSet<ChangeHash>> for JS {
-    fn from(heads: HashSet<ChangeHash>) -> Self {
-        let result: JsValue = Object::new().into();
-        for key in &heads {
-            Reflect::set(&result, &key.to_string().into(), &true.into()).unwrap();
-        }
-        JS(result)
-    }
-}
-
-impl From<Option<Vec<ChangeHash>>> for JS {
-    fn from(heads: Option<Vec<ChangeHash>>) -> Self {
-        if let Some(v) = heads {
-            let v: Array = v
-                .iter()
-                .map(|h| JsValue::from_str(&h.to_string()))
-                .collect();
-            JS(v.into())
-        } else {
-            JS(JsValue::null())
-        }
-    }
-}
-
-impl TryFrom<JS> for HashSet<ChangeHash> {
-    type Error = JsValue;
-
-    fn try_from(value: JS) -> Result<Self, Self::Error> {
-        let mut result = HashSet::new();
-        for key in Reflect::own_keys(&value.0)?.iter() {
-            if let Some(true) = Reflect::get(&value.0, &key)?.as_bool() {
-                result.insert(key.into_serde().map_err(to_js_err)?);
-            }
-        }
-        Ok(result)
-    }
-}
-
-impl TryFrom<JS> for Vec<ChangeHash> {
-    type Error = JsValue;
-
-    fn try_from(value: JS) -> Result<Self, Self::Error> {
-        let value = value.0.dyn_into::<Array>()?;
-        let value: Result<Vec<ChangeHash>, _> = value.iter().map(|j| j.into_serde()).collect();
-        let value = value.map_err(to_js_err)?;
-        Ok(value)
-    }
-}
-
-impl From<JS> for Option<Vec<ChangeHash>> {
-    fn from(value: JS) -> Self {
-        let value = value.0.dyn_into::<Array>().ok()?;
-        let value: Result<Vec<ChangeHash>, _> = value.iter().map(|j| j.into_serde()).collect();
-        let value = value.ok()?;
-        Some(value)
-    }
-}
-
-impl TryFrom<JS> for Vec<Change> {
-    type Error = JsValue;
-
-    fn try_from(value: JS) -> Result<Self, Self::Error> {
-        let value = value.0.dyn_into::<Array>()?;
-        let changes: Result<Vec<Uint8Array>, _> = value.iter().map(|j| j.dyn_into()).collect();
-        let changes = changes?;
-        let changes: Result<Vec<Change>, _> = changes
-            .iter()
-            .map(|a| am::decode_change(a.to_vec()))
-            .collect();
-        let changes = changes.map_err(to_js_err)?;
-        Ok(changes)
-    }
-}
-
-impl TryFrom<JS> for am::SyncState {
-    type Error = JsValue;
-
-    fn try_from(value: JS) -> Result<Self, Self::Error> {
-        let value = value.0;
-        let shared_heads = get(&value, "sharedHeads")?.try_into()?;
-        let last_sent_heads = get(&value, "lastSentHeads")?.try_into()?;
-        let their_heads = get(&value, "theirHeads")?.into();
-        let their_need = get(&value, "theirNeed")?.into();
-        let their_have = get(&value, "theirHave")?.try_into()?;
-        let sent_hashes = get(&value, "sentHashes")?.try_into()?;
-        Ok(am::SyncState {
-            shared_heads,
-            last_sent_heads,
-            their_heads,
-            their_need,
-            their_have,
-            sent_hashes,
-        })
-    }
-}
-
-impl TryFrom<JS> for Option<Vec<am::SyncHave>> {
-    type Error = JsValue;
-
-    fn try_from(value: JS) -> Result<Self, Self::Error> {
-        if value.0.is_null() {
-            Ok(None)
-        } else {
-            Ok(Some(value.try_into()?))
-        }
-    }
-}
-
-impl TryFrom<JS> for Vec<am::SyncHave> {
-    type Error = JsValue;
-
-    fn try_from(value: JS) -> Result<Self, Self::Error> {
-        let value = value.0.dyn_into::<Array>()?;
-        let have: Result<Vec<am::SyncHave>, JsValue> = value
-            .iter()
-            .map(|s| {
-                let last_sync = get(&s, "lastSync")?.try_into()?;
-                let bloom = get(&s, "bloom")?.try_into()?;
-                Ok(am::SyncHave { last_sync, bloom })
-            })
-            .collect();
-        let have = have?;
-        Ok(have)
-    }
-}
-
-impl TryFrom<JS> for am::BloomFilter {
-    type Error = JsValue;
-
-    fn try_from(value: JS) -> Result<Self, Self::Error> {
-        let value: Uint8Array = value.0.dyn_into()?;
-        let value = value.to_vec();
-        let value = value.as_slice().try_into().map_err(to_js_err)?;
-        Ok(value)
-    }
-}
-
-impl From<&[ChangeHash]> for AR {
-    fn from(value: &[ChangeHash]) -> Self {
-        AR(value
-            .iter()
-            .map(|h| JsValue::from_str(&hex::encode(&h.0)))
-            .collect())
-    }
-}
-
-impl From<&[Change]> for AR {
-    fn from(value: &[Change]) -> Self {
-        let changes: Array = value
-            .iter()
-            .map(|c| Uint8Array::from(c.raw_bytes()))
-            .collect();
-        AR(changes)
-    }
-}
-
-impl From<&[am::SyncHave]> for AR {
-    fn from(value: &[am::SyncHave]) -> Self {
-        AR(value
-            .iter()
-            .map(|have| {
-                let last_sync: Array = have
-                    .last_sync
-                    .iter()
-                    .map(|h| JsValue::from_str(&hex::encode(&h.0)))
-                    .collect();
-                // FIXME - the clone and the unwrap here shouldnt be needed - look at into_bytes()
-                let bloom = Uint8Array::from(have.bloom.clone().into_bytes().unwrap().as_slice());
-                let obj: JsValue = Object::new().into();
-                // we can unwrap here b/c we created the object and know its not frozen
-                Reflect::set(&obj, &"lastSync".into(), &last_sync.into()).unwrap();
-                Reflect::set(&obj, &"bloom".into(), &bloom.into()).unwrap();
-                obj
-            })
-            .collect())
-    }
-}
-
-fn get_heads(heads: JsValue) -> Option<Vec<ChangeHash>> {
-    JS(heads).into()
-}
-
-fn map_to_js(doc: &mut am::Automerge, obj: &ObjId) -> JsValue {
-    let keys = doc.keys(obj);
-    let map = Object::new();
-    for k in keys {
-        let val = doc.value(obj, &k);
-        match val {
-            Ok(Some((Value::Object(o), exid)))
-                if o == am::ObjType::Map || o == am::ObjType::Table =>
-            {
-                Reflect::set(&map, &k.into(), &map_to_js(doc, &exid)).unwrap();
-            }
-            Ok(Some((Value::Object(_), exid))) => {
-                Reflect::set(&map, &k.into(), &list_to_js(doc, &exid)).unwrap();
-            }
-            Ok(Some((Value::Scalar(v), _))) => {
-                Reflect::set(&map, &k.into(), &ScalarValue(v).into()).unwrap();
-            }
-            _ => (),
-        };
-    }
-    map.into()
-}
-
-fn list_to_js(doc: &mut am::Automerge, obj: &ObjId) -> JsValue {
-    let len = doc.length(obj);
-    let array = Array::new();
-    for i in 0..len {
-        let val = doc.value(obj, i as usize);
-        match val {
-            Ok(Some((Value::Object(o), exid)))
-                if o == am::ObjType::Map || o == am::ObjType::Table =>
-            {
-                array.push(&map_to_js(doc, &exid));
-            }
-            Ok(Some((Value::Object(_), exid))) => {
-                array.push(&list_to_js(doc, &exid));
-            }
-            Ok(Some((Value::Scalar(v), _))) => {
-                array.push(&ScalarValue(v).into());
-            }
-            _ => (),
-        };
-    }
-    array.into()
-}
